requirements.txt
streamlit
opencv-python
numpy
torch
ultralytics
supervision
sort-tracker
torchvision
